@@ -1,94 +1,71 @@
 {
-  "name": "ionic-hello-world",
-  "version": "0.0.0",
+  "name": "i-diario-ap-v2",
+  "version": "0.0.1",
   "author": "Ionic Framework",
-  "homepage": "http://ionicframework.com/",
-  "private": true,
+  "homepage": "https://ionicframework.com/",
   "scripts": {
-    "clean": "ionic-app-scripts clean",
-    "build": "ionic-app-scripts build --release --prod --aot",
-    "lint": "ionic-app-scripts lint",
-    "ionic:build": "ionic-app-scripts build",
-    "ionic:serve": "ionic-app-scripts serve"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "lint": "ng lint"
   },
-  "config": {
-    "ionic_generate_source_map": "true"
-  },
+  "private": true,
   "dependencies": {
-    "@angular/common": "5.2.10",
-    "@angular/compiler": "5.2.10",
-    "@angular/compiler-cli": "5.2.10",
-    "@angular/core": "5.2.10",
-    "@angular/forms": "5.2.10",
-    "@angular/http": "5.2.10",
-    "@angular/platform-browser": "5.2.10",
-    "@angular/platform-browser-dynamic": "5.2.10",
-    "@ionic-native/core": "^4.7.0",
-    "@ionic-native/device": "^4.7.0",
-    "@ionic-native/file": "^4.7.0",
-    "@ionic-native/in-app-browser": "^4.7.0",
-    "@ionic-native/network": "^4.7.0",
-    "@ionic-native/safari-view-controller": "^4.7.0",
-    "@ionic-native/splash-screen": "4.7.0",
-    "@ionic-native/status-bar": "4.7.0",
-    "@ionic/storage": "^2.1.3",
-    "cordova": "^8.0.0",
-    "cordova-android": "7.1.0",
-    "cordova-browser": "5.0.3",
-    "cordova-ios": "5.1.1",
-    "cordova-plugin-console": "^1.1.0",
-    "cordova-plugin-device": "^2.0.3",
-    "cordova-plugin-file": "^6.0.2",
-    "cordova-plugin-inappbrowser": "^4.1.0",
-    "cordova-plugin-ionic": "^5.5.1",
-    "cordova-plugin-network-information": "^2.0.2",
-    "cordova-plugin-safariviewcontroller": "^1.6.0",
-    "cordova-plugin-splashscreen": "^5.0.4",
-    "cordova-plugin-statusbar": "^2.4.3",
-    "cordova-plugin-whitelist": "^1.3.4",
-    "cordova-sqlite-storage": "^2.6.0",
-    "core-js": "^2.5.7",
-    "ionic-angular": "3.9.2",
-    "ionic-plugin-keyboard": "^2.2.1",
-    "ionicons": "3.0.0",
-    "node-sass": "^4.12.0",
-    "rxjs": "5.5.10",
-    "sw-toolbox": "3.6.0",
-    "unorm": "^1.4.1",
-    "zone.js": "0.8.26"
+    "@angular/animations": "^17.0.2",
+    "@angular/common": "^17.0.2",
+    "@angular/compiler": "^17.0.2",
+    "@angular/core": "^17.0.2",
+    "@angular/forms": "^17.0.2",
+    "@angular/platform-browser": "^17.0.2",
+    "@angular/platform-browser-dynamic": "^17.0.2",
+    "@angular/router": "^17.0.2",
+    "@capacitor/android": "^6.1.1",
+    "@capacitor/app": "^5.0.8",
+    "@capacitor/browser": "^6.0.0",
+    "@capacitor/core": "^6.1.1",
+    "@capacitor/device": "^6.0.0",
+    "@capacitor/haptics": "5.0.7",
+    "@capacitor/ios": "^6.1.1",
+    "@capacitor/keyboard": "5.0.8",
+    "@capacitor/network": "^5.0.8",
+    "@capacitor/status-bar": "^5.0.8",
+    "@ionic/angular": "^7.0.0",
+    "@ionic/storage-angular": "^4.0.0",
+    "ionicons": "^7.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@ionic/app-scripts": "3.2.4",
-    "ionic": "3.20.0",
-    "typescript": "^2.8.1"
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular-eslint/builder": "^17.0.0",
+    "@angular-eslint/eslint-plugin": "^17.0.0",
+    "@angular-eslint/eslint-plugin-template": "^17.0.0",
+    "@angular-eslint/schematics": "^17.0.0",
+    "@angular-eslint/template-parser": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.2",
+    "@angular/language-service": "^17.0.2",
+    "@capacitor/assets": "^3.0.5",
+    "@capacitor/cli": "^5.7.0",
+    "@ionic/angular-toolkit": "^9.0.0",
+    "@types/jasmine": "~5.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "eslint": "^7.26.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
+    "jasmine-core": "~5.1.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.2.2"
   },
-  "cordovaPlugins": [
-    "cordova-plugin-console",
-    "cordova-plugin-whitelist",
-    "cordova-plugin-device",
-    "cordova-plugin-splashscreen",
-    "cordova-plugin-statusbar",
-    "ionic-plugin-keyboard"
-  ],
-  "cordovaPlatforms": [],
-  "description": "app_offline: An Ionic project",
-  "cordova": {
-    "plugins": {
-      "cordova-plugin-network-information": {},
-      "cordova-plugin-console": {},
-      "cordova-plugin-device": {},
-      "cordova-plugin-splashscreen": {},
-      "cordova-plugin-statusbar": {},
-      "cordova-plugin-whitelist": {},
-      "ionic-plugin-keyboard": {},
-      "cordova-plugin-inappbrowser": {},
-      "cordova-sqlite-storage": {},
-      "cordova-plugin-safariviewcontroller": {},
-      "cordova-plugin-file": {}
-    },
-    "platforms": [
-      "android",
-      "ios"
-    ]
-  }
+  "description": "An Ionic project"
 }
